fix ProductDTO and product request object to work with stock

+++ src/main/java/com/example/techcentral/models/Product.java
@@ -27,7 +27,7 @@ public class Product {
     private double price;
 
     @Column(name = "PRODUCT_STOCK")
-    private double stock;
+    private int stock;
 
     @OneToMany(mappedBy = "product", cascade = CascadeType.ALL, fetch = FetchType.EAGER)
     private List<ProductImage> productImages;
